+++ html_stream/render_html_body.ts
@@ -1,6 +1,7 @@
 import { prependDocType } from "./transform/prepend_doctype.ts";
 import { isSafe } from "./token.ts";
 import { safetyFilter } from "@http/token-stream/transform/safety-filter";
+import { catchErrors } from "@http/token-stream/transform/catch-errors";
 import { renderBody } from "@http/token-stream/render-body";
 import type { HtmlNode, HtmlToken } from "./types.ts";
 import type { RenderOptions } from "@http/token-stream/types";
@@ -21,6 +22,7 @@ export function renderHtmlBody(
     prependDocType(),
     ...(options?.transformers ?? []),
     safetyFilter(isSafe),
+    catchErrors<HtmlToken>(),
   ];
   return renderBody(node, {
     ...options,